Make schema builder an extendable class
